+++ src/python_1c_odata/chart_of_characteristic_types.py
@@ -1,0 +1,5 @@
+from python_1c_odata.entity import EntitySet
+
+
+class ChartOfCharacteristicTypes(EntitySet):
+    kind = "ChartOfCharacteristicTypes"